# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lander Dispatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loads:
    get:
      tags: [Loads]
      operationId: listLoads
      summary: List loads
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: carrierId
          in: query
          required: false
          schema:
            type: string
        - name: brokerId
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated load list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadListResponse'
    post:
      tags: [Loads]
      operationId: createLoad
      summary: Create a load
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
  /loads/{loadId}:
    get:
      tags: [Loads]
      operationId: getLoad
      summary: Get a load by ID
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Load detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
    patch:
      tags: [Loads]
      operationId: updateLoad
      summary: Update a load
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
    delete:
      tags: [Loads]
      operationId: deleteLoad
      summary: Delete a load
      parameters:
        - name: loadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  # CRM - Contacts
...
  /invoices:
    get:
      tags: [Invoices]
      operationId: listInvoices
      summary: List invoices
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: carrierId
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceListResponse'
    post:
      tags: [Invoices]
      operationId: createInvoice
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
  /invoices/{invoiceId}:
    get:
      tags: [Invoices]
      operationId: getInvoice
      summary: Get an invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
    patch:
      tags: [Invoices]
      operationId: updateInvoice
      summary: Update an invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
    delete:
      tags: [Invoices]
      operationId: deleteInvoice
      summary: Delete an invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /invoices/{invoiceId}/payments:
    post:
      tags: [Invoices]
      operationId: recordInvoicePayment
      summary: Record a payment for an invoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoicePaymentInput'
      responses:
        '201':
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'

  # Transactions
...
components:
...
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # Dashboard
...
    LoadVehicle:
      type: object
      required: [vehicleNumber]
      properties:
        vehicleNumber:
          type: integer
        year:
          type: [string, 'null']
        make:
          type: [string, 'null']
        model:
          type: [string, 'null']
        type:
          type: [string, 'null']
        color:
          type: [string, 'null']
        plate:
          type: [string, 'null']
        vin:
          type: [string, 'null']
        lotNumber:
          type: [string, 'null']
        additionalInfo:
          type: [string, 'null']
    Load:
      type: object
      required: [id, loadId, status, createdAt]
      properties:
        id:
          type: string
        loadId:
          type: string
        carrierId:
          type: [string, 'null']
        carrierName:
          type: [string, 'null']
        brokerId:
          type: [string, 'null']
        brokerName:
          type: [string, 'null']
        status:
          type: string
        dispatchDate:
          type: [string, 'null']
          format: date
        pickupCity:
          type: [string, 'null']
        pickupState:
          type: [string, 'null']
        pickupZip:
          type: [string, 'null']
        pickupEstimated:
          type: [string, 'null']
          format: date
        pickupDeadline:
          type: [string, 'null']
          format: date
        deliveryCity:
          type: [string, 'null']
        deliveryState:
          type: [string, 'null']
        deliveryZip:
          type: [string, 'null']
        deliveryEstimated:
          type: [string, 'null']
          format: date
        deliveryDeadline:
          type: [string, 'null']
          format: date
        miles:
          type: [number, 'null']
        rate:
          type: [number, 'null']
        ratePerMile:
          type: [number, 'null']
        paymentMethod:
          type: [string, 'null']
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/LoadVehicle'
        dispatchInstructions:
          type: [string, 'null']
        pickupInstructions:
          type: [string, 'null']
        deliveryInstructions:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    LoadInput:
      type: object
      properties:
        carrierId:
          type: string
        brokerId:
          type: string
        status:
          type: string
        dispatchDate:
          type: string
        pickupCity:
          type: string
        pickupState:
          type: string
        pickupZip:
          type: string
        pickupEstimated:
          type: string
        pickupDeadline:
          type: string
        deliveryCity:
          type: string
        deliveryState:
          type: string
        deliveryZip:
          type: string
        deliveryEstimated:
          type: string
        deliveryDeadline:
          type: string
        miles:
          type: number
        rate:
          type: number
        paymentMethod:
          type: string
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/LoadVehicle'
        dispatchInstructions:
          type: string
        pickupInstructions:
          type: string
        deliveryInstructions:
          type: string
    LoadUpdate:
      type: object
      properties:
        carrierId:
          type: string
        brokerId:
          type: string
        status:
          type: string
        dispatchDate:
          type: string
        pickupCity:
          type: string
        pickupState:
          type: string
        pickupZip:
          type: string
        pickupEstimated:
          type: string
        pickupDeadline:
          type: string
        deliveryCity:
          type: string
        deliveryState:
          type: string
        deliveryZip:
          type: string
        deliveryEstimated:
          type: string
        deliveryDeadline:
          type: string
        miles:
          type: number
        rate:
          type: number
        paymentMethod:
          type: string
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/LoadVehicle'
        dispatchInstructions:
          type: string
        pickupInstructions:
          type: string
        deliveryInstructions:
          type: string
    LoadListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Load'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # CRM Contact
...
    InvoicePayment:
      type: object
      required: [id, paymentDate, amount, paymentMethod]
      properties:
        id:
          type: string
        paymentDate:
          type: string
          format: date
        amount:
          type: number
        paymentMethod:
          type: string
        reference:
          type: [string, 'null']
        notes:
          type: [string, 'null']
    Invoice:
      type: object
      required: [id, invoiceNumber, status, total, amountPaid, balance, createdAt]
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        carrierId:
          type: [string, 'null']
        carrierName:
          type: [string, 'null']
        issueDate:
          type: [string, 'null']
          format: date
        dueDate:
          type: [string, 'null']
          format: date
        total:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        status:
          type: string
        notes:
          type: [string, 'null']
        payments:
          type: array
          items:
            $ref: '#/components/schemas/InvoicePayment'
        loadIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    InvoiceInput:
      type: object
      required: [carrierId, total]
      properties:
        carrierId:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        total:
          type: number
        loadIds:
          type: array
          items:
            type: string
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        carrierId:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        total:
          type: number
        loadIds:
          type: array
          items:
            type: string
        notes:
          type: string
        status:
          type: string
    InvoicePaymentInput:
      type: object
      required: [paymentDate, amount, paymentMethod]
      properties:
        paymentDate:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
        reference:
          type: string
        notes:
          type: string
    InvoiceListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Transaction
...